Separated out various implementations of RasterSource into their own files.
