+++ core/src/main/resources/reference.conf
@@ -3,6 +3,7 @@ rasterframes {
   prefer-gdal = true
   showable-tiles = true
   showable-max-cells = 20
+  raster-source-cache-timeout = 120 seconds
 }
 
 vlm.gdal {
